[Feat #265] useClickHandler.ts 생성
